requirements/test.txt
-r base.txt

pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.2.0
Faker>=18.0.0
